move no param test to main app
