+++ dev/io.openliberty.mcp.internal_fat/fat/src/io/openliberty/mcp/internal/fat/FATSuite.java
@@ -25,6 +25,7 @@ import io.openliberty.mcp.internal.fat.protocol.HttpTest;
 import io.openliberty.mcp.internal.fat.protocol.ProtocolVersionTest;
 import io.openliberty.mcp.internal.fat.tool.CancellationTest;
 import io.openliberty.mcp.internal.fat.tool.DeploymentProblemTest;
+import io.openliberty.mcp.internal.fat.tool.NoParamNameTest;
 import io.openliberty.mcp.internal.fat.tool.ToolErrorHandlingTest;
 import io.openliberty.mcp.internal.fat.tool.ToolTest;
 
@@ -38,6 +39,7 @@ import io.openliberty.mcp.internal.fat.tool.ToolTest;
                 CancellationTest.class,
                 HttpTest.class,
                 LifecycleTest.class,
+                NoParamNameTest.class,
                 ProtocolVersionTest.class,
                 ToolErrorHandlingTest.class,
                 ToolTest.class
