+++ dev/io.openliberty.mcp.internal_fat/fat/src/io/openliberty/mcp/internal/fat/tool/NoParamNameTest.java
@@ -7,10 +7,9 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package io.openliberty.mcp.internal.fat.noparamtool;
+package io.openliberty.mcp.internal.fat.tool;
 
 import static com.ibm.websphere.simplicity.ShrinkHelper.DeployOptions.SERVER_ONLY;
-import static org.junit.Assert.assertNotNull;
 
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
@@ -20,7 +19,6 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.skyscreamer.jsonassert.JSONAssert;
-import org.skyscreamer.jsonassert.JSONCompareMode;
 
 import com.ibm.websphere.simplicity.ShrinkHelper;
 
@@ -28,31 +26,28 @@ import componenttest.annotation.Server;
 import componenttest.custom.junit.runner.FATRunner;
 import componenttest.topology.impl.LibertyServer;
 import componenttest.topology.utils.FATServletClient;
-import io.openliberty.mcp.internal.fat.noparamtool.basicToolApp.BasicTools;
+import io.openliberty.mcp.internal.fat.noparamtool.NoParamTools;
 import io.openliberty.mcp.internal.fat.utils.HttpTestUtils;
 
-/**
- *
- */
 @RunWith(FATRunner.class)
-public class ToolTest extends FATServletClient {
+public class NoParamNameTest extends FATServletClient {
     @Server("mcp-server")
     public static LibertyServer server;
 
     @BeforeClass
     public static void setup() throws Exception {
-        WebArchive war = ShrinkWrap.create(WebArchive.class, "toolTest.war").addPackage(BasicTools.class.getPackage());
+        WebArchive war = ShrinkWrap.create(WebArchive.class, "toolTest.war").addPackage(NoParamTools.class.getPackage());
 
         ShrinkHelper.exportDropinAppToServer(server, war, SERVER_ONLY);
 
         server.startServer();
     }
-    
+
     @AfterClass
     public static void teardown() throws Exception {
         server.stopServer();
     }
-    
+
     @Test
     public void testIllegalToolArgNameTool() throws Exception {
         String request = """
@@ -73,7 +68,7 @@ public class ToolTest extends FATServletClient {
         JSONObject jsonResponse = new JSONObject(response);
 
         String expectedResponseString = """
-                        {"id":\"2\","jsonrpc":"2.0","result":{"content":[{"type":"text","text":"Hello"}], "isError": false}}
+                        Exception thrown
                         """;
         JSONAssert.assertEquals(expectedResponseString, response, true);
     }
